feat(): added dashboard feature for clean api for all metrics

+++ src/main/java/com/app/budgets/budget/dto/RecurringMetrics.java
@@ -3,6 +3,6 @@ package com.app.budgets.budget.dto;
 import java.math.BigDecimal;
 
 public interface RecurringMetrics {
-    BigDecimal getCount();
+    Integer getCount();
     BigDecimal getSum();
 }
