+++ src/main/java/com/app/budgets/dashboard/dto/ReminderMetrics.java
@@ -3,8 +3,8 @@ package com.app.budgets.dashboard.dto;
 import java.time.LocalDate;
 
 public interface ReminderMetrics {
-    Long getOverdue();      // Overdue count
-    Long getDueSoon();      // Due in next 7 days
-    Long getTotal();        // Total active reminders
+    Integer getOverdue();      // Overdue count
+    Integer getDueSoon();      // Due in next 7 days
+    Integer getTotal();        // Total active reminders
     LocalDate getNextDueDate();
 }
